change weight to double. it had integer overflow. do I even need it anymore?

+++ EMCal_pi0_Calib_2023/src/CaloAna.h
@@ -182,7 +182,7 @@ class CaloAna : public SubsysReco
   TH1F* h_truth_pid_cuts[6];
 
   float target_pi0_mass = 0.145;
-  int WeightScale=1e+14;
+  double WeightScale=1e+14;
   double inv_yield;
 
 };
